Fix double reload for most edits by moving ClientDeploymentConfiguration location
