+++ packages/types/src/deployment-configuration.ts
@@ -1,0 +1,23 @@
+/**
+ * @license
+ * Copyright 2025 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { type RuntimeFlags } from "./flags.js";
+
+export type ClientDeploymentConfiguration = {
+  MEASUREMENT_ID?: string;
+  BACKEND_API_ENDPOINT?: string;
+  ENABLE_GOOGLE_DRIVE_PROXY?: boolean;
+  FEEDBACK_LINK?: string;
+  ENABLE_GOOGLE_FEEDBACK?: boolean;
+  GOOGLE_FEEDBACK_PRODUCT_ID?: string;
+  GOOGLE_FEEDBACK_BUCKET?: string;
+  flags: RuntimeFlags;
+};
+
+export type ServerDeploymentConfiguration = {
+  BACKEND_API_ENDPOINT?: string;
+  ENABLE_GOOGLE_DRIVE_PROXY?: boolean;
+};
