+++ packages/unified-server/src/server/provide-config.ts
@@ -5,20 +5,12 @@
  */
 
 import { SecretsProvider } from "@breadboard-ai/board-server";
-import { type ClientDeploymentConfiguration } from "@breadboard-ai/shared-ui/config/client-deployment-configuration.js";
+import {
+  type ClientDeploymentConfiguration,
+  type ServerDeploymentConfiguration,
+} from "@breadboard-ai/types/deployment-configuration.js";
 
 export { getConfigFromSecretManager };
-
-/**
- * Server-specific variables of deployment configuration. These values
- * are only available to the server. To add a new value:
- * 1. Add it here
- * 2. Consume it in `./main.ts`
- */
-export type ServerDeploymentConfiguration = {
-  BACKEND_API_ENDPOINT?: string;
-  ENABLE_GOOGLE_DRIVE_PROXY?: boolean;
-};
 
 export type SecretValueFormat = {
   client: ClientDeploymentConfiguration;
